+++ ftpsync/synchronizers.py
@@ -185,6 +185,11 @@ class BaseSynchronizer(object):
                     self.remote.get_base_name(),
                 )
             )
+            write(
+                "Encoding local: {}, remote: {}".format(
+                    self.local.encoding, self.remote.encoding
+                )
+            )
 
         try:
             self.local.synchronizer = self.remote.synchronizer = self
@@ -516,6 +521,10 @@ class BaseSynchronizer(object):
             if remote_entry.name not in local_entry_map:
                 entry_pair = EntryPair(None, remote_entry)
                 entry_pair_list.append(entry_pair)
+                # print("NOT IN LOCAL")
+                # print(remote_entry.name)
+                # print(local_entry_map.keys())
+                # print(self.local.cur_dir_meta.peer_sync.get(self.remote.get_id()))
 
         # 3. Classify all entries and pairs.
         #    We pass the additional meta data here